Contact List Application
A Java application that allows the user to enter a series of contacts’
names, ages, e-mail addresses, and cell phone numbers, then creates a
file from the entered data while validating the age entry to ensure that
it is numeric and between 0 and 120. Also allows the user to view the
records within the application.